fix: test corrections — overlay assertion, enforcement naming, CLAUDE.md
Update test_disable_quota_guard_writes_disabled_flag to assert overlay
file instead of base. Move session-scope tests to test_session_start_hook.py
matching the enforcement naming convention. Add entry to hooks CLAUDE.md.
