+++ tests/hooks/CLAUDE.md
@@ -22,6 +22,7 @@ Hook script behavior, registration, and bridge tests.
 | `test_recipe_write_advisor.py` | Tests for autoskillit.hooks.guards.recipe_write_advisor |
 | `test_review_gate_post_hook.py` | Tests for the review_gate_post_hook PostToolUse hook |
 | `test_review_loop_gate.py` | Tests for the review_loop_gate PreToolUse hook |
+| `test_session_start_hook.py` | Session-scope enforcement tests for session_start_hook.py |
 | `test_session_start_reminder.py` | Tests for the SessionStart hook — session_start_hook.py |
 | `test_skill_load_post_hook.py` | Tests for skill_load_post_hook.py PostToolUse hook |
 | `test_token_summary_appender.py` | Tests for the token_summary_appender PostToolUse hook |
