+++ tests/hooks/test_session_start_reminder.py
@@ -291,47 +291,3 @@ def test_session_start_hook_resolves_campaign_namespace(tmp_path: Path, monkeypa
     # The hook should find the marker and produce a resume reminder
     data = json.loads(out.strip())
     assert "additionalContext" in data
-
-
-# --- Group S: session_scope = interactive_only ---
-
-
-def test_session_start_hook_silent_for_headless_sessions(tmp_path: Path) -> None:
-    """Headless sessions must not receive the open-kitchen reminder."""
-    transcript = tmp_path / "session.jsonl"
-    transcript.write_text('{"type":"say","text":"hello"}\n')
-    payload = json.dumps({"session_id": "abc", "transcript_path": str(transcript)})
-    env = {
-        **os.environ,
-        "AUTOSKILLIT_HEADLESS": "1",
-        "AUTOSKILLIT_STATE_DIR": str(tmp_path / "state"),
-    }
-    rc, out = _run(payload, env=env)
-    assert rc == 0
-    assert "additionalContext" not in out
-
-
-def test_session_start_hook_fires_for_interactive_sessions(tmp_path: Path) -> None:
-    """Non-headless resumed sessions must still receive the open-kitchen reminder."""
-    transcript = tmp_path / "session.jsonl"
-    transcript.write_text('{"type":"say","text":"hello"}\n')
-    payload = json.dumps({"session_id": "abc", "transcript_path": str(transcript)})
-    env = {k: v for k, v in os.environ.items() if k != "AUTOSKILLIT_HEADLESS"}
-    env["AUTOSKILLIT_STATE_DIR"] = str(tmp_path / "empty_state")
-    rc, out = _run(payload, env=env)
-    assert rc == 0
-    data = json.loads(out.strip())
-    assert "additionalContext" in data
-    assert "open-kitchen" in data["additionalContext"]
-
-
-def test_session_start_hook_registry_scope() -> None:
-    """session_start_hook must be registered with session_scope=interactive_only."""
-    from autoskillit.hook_registry import HOOK_REGISTRY
-
-    hook_def = next(
-        h
-        for h in HOOK_REGISTRY
-        if h.event_type == "SessionStart" and "session_start_hook.py" in h.scripts
-    )
-    assert hook_def.session_scope == "interactive_only"
